+++ CHANGELOG.md
@@ -202,5 +202,5 @@ Complete rewrite of the application stack for better performance, type safety, a
 - Remove Spacy and use a simpler tokenizer instead,lighter as well
 - add subliminal as a priority source for words as srt for movies
 - add nested pagination in levels of words, combine c1 and c2 together only on ui until we figure out why c1 doesnt get filled
-
+- add translation from DeepL Free tier, (azerbaijani is not supported as a target language)